+++ src/components/codeblocks/CodeblockTree.svelte
@@ -16,6 +16,7 @@
 	import CopyToClipboardButton from "../button/CopyToClipboardButton.svelte";
 	import CodeblockErrors from "./CodeblockErrors.svelte";
 	import { to_node_stringify_options } from "src/graph/utils";
+	import { useOwned } from "src/stores/use_owned.svelte";
 
 	interface Props {
 		plugin: BreadcrumbsPlugin;
@@ -26,16 +27,13 @@
 
 	let { plugin, options, errors, file_path }: Props = $props();
 
-	let node_stringify_options = $derived(
+	const owned_stringify = useOwned(() =>
 		to_node_stringify_options(
 			plugin.settings,
 			plugin.settings.views.codeblocks.show_node_options,
 		),
 	);
-	$effect(() => {
-		const o = node_stringify_options;
-		return () => o.free();
-	});
+	let node_stringify_options = $derived(owned_stringify.current);
 
 	const DEFAULT_MAX_DEPTH = 5;